Change Select menu entry text for selected windows.
Other toggle menu items such as Maximize and Shade change their text
labels according to the action which will be performed.  A shaded window
has its Shade menu item text changed to Unshade, for instance.

As well as maintaining consistency with other menu items, changing the
Select menu entry's text to Deselect for current selected windows
provides another (the only) way of reporting that a window is selected
when its border has been disabled.  Currently the only indication that a
window is selected is that its border colour will change.