Updated default icons
All the non-Window Maker-specific icons have been removed from the default
configuration in WMWindowAttributes.

In addition, the icon for drawers has been added to WMWindowAttributes.  Also,
it has been changed from a .png file to a .tiff and .xpm to bring it in line
with the other provided icons.

Update:  in the original version of this patch, I neglected to modify the
appropriate Makefile.am to account for the drawer icon changes.